Move sidebar files into /components subfolders
